refactor(architecture): updating
