+++ .vscode/cSpell.json
@@ -7,7 +7,8 @@
     "codecov",
     "webhook",
     "rocu",
-    "promisify"
+    "promisify",
+    "portfinder"
   ],
   "flagWords": [
       "hte"
