+++ src/cli.ts
@@ -7,13 +7,13 @@ import opn = require("opn");
 import * as path from "path";
 import * as readPkgUp from "read-pkg-up";
 import { UpdateNotifier } from "update-notifier";
-import { log } from "../logger";
+import { log } from "./logger";
 
-const pkg = require("../../package.json");
+const pkg = require("../package.json");
 new UpdateNotifier({ pkg }).notify();
 
-import { server } from "../server";
-import { generateStaticPage } from "./generateStatic";
+import { generateStaticPage } from "./generator";
+import { server } from "./server";
 
 const cli = meow(
   `
