+++ src/index.ts
@@ -1,2 +1,5 @@
-export * from "./cli";
-export { generateStatic } from "./generator";
+export { generateStaticPage } from "./generator";
+export { log } from "./logger";
+export { render } from "./renderer";
+export { server } from "./server";
+export { getData } from "./structure";
