+++ src/logger/index.ts
@@ -1,5 +1,5 @@
 import chalk from "chalk";
 
 export const log = (...messages: any[]) => {
-  console.log(chalk.black.bgCyan(" rocu "), chalk.cyan(...messages));
+  console.log(chalk.black.bgRedBright(" rocu "), chalk.redBright(...messages));
 };
