+++ src/renderer/index.ts
@@ -1,16 +1,14 @@
 import { Options } from "@rocu/cli";
-import { PageElement, RenderedPage, Source } from "@rocu/page";
+import { PageElement, RenderedStaticPage, Source } from "@rocu/page";
 
-const renderPage = (page: PageElement): RenderedPage => {
+const renderPage = (page: PageElement): RenderedStaticPage => {
   return {
-    ...page,
+    name: page.name,
     html: "<body>Hello</body>",
   };
 };
 
-const render = async ({ dirname, pages = [] }: Source, opts: Options): Promise<RenderedPage[]> => {
-  console.log(dirname);
-  console.log(pages);
+const render = async ({ dirname, pages = [] }: Source, opts: Options): Promise<RenderedStaticPage[]> => {
   return pages.map(renderPage);
 };
