+++ src/server/index.ts
@@ -1,2 +1,1 @@
 export { server } from "./server";
-export { getData } from "./getPage";
