+++ src/structure/index.ts
@@ -1,0 +1,1 @@
+export { getData } from "./getPage";
